Move connect-mixin to components/core
