+++ src/components/core/connect-mixin.ts
@@ -2,9 +2,9 @@
 
 import { AnyAction, ThunkAction, Unsubscribe } from '@reduxjs/toolkit';
 import { property } from 'lit/decorators.js';
-import { AppStore, RootState, store as globalStore } from '../store.js';
-import { Constructor } from '../util/shared-types.js';
-import { SliceConfigurator } from './slice-configurator.js';
+import { AppStore, RootState, store as globalStore } from '../../app/store.js';
+import { SliceConfigurator } from '../../middleware/slice-configurator.js';
+import { Constructor } from '../../util/shared-types.js';
 
 interface CustomElement {
   connectedCallback?(): void;
