+++ src/components/lineup-view-game-detail.ts
@@ -5,9 +5,9 @@ import { html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { updateMetadata } from 'pwa-helpers/metadata.js';
 import { RootState } from '../app/store.js';
-import { ConnectStoreMixin } from '../middleware/connect-mixin.js';
 import { GameDetail, GameStatus } from '../models/game.js';
 import { getGame, getGameSliceConfigurator, selectGameById } from '../slices/game/index.js';
+import { ConnectStoreMixin } from './core/connect-mixin.js';
 import './lineup-game-complete.js';
 import './lineup-game-live.js';
 import './lineup-game-setup.js';
